+++ src/slice-assign.test.ts
@@ -5,7 +5,7 @@ describe('assignSlice', () => {
     it('should set state to payload', () => {
       const name = 'test';
       const { reducer, actions } = assignSlice({ name, initialState: 0 });
-      const actual = reducer(0, actions.setTest(2));
+      const actual = reducer(0, actions.set(2));
       expect(actual).toEqual(2);
     });
   });
@@ -14,7 +14,7 @@ describe('assignSlice', () => {
     it('should reset to initialState', () => {
       const name = 'test';
       const { reducer, actions } = assignSlice({ name, initialState: 5 });
-      const actual = reducer(0, actions.resetTest(2));
+      const actual = reducer(0, actions.reset());
       expect(actual).toEqual(5);
     });
   });