+++ src/util.ts
@@ -1,1 +1,0 @@
-export const cap = (t: string) => t.charAt(0).toUpperCase() + t.substr(1);