+++ type-tests/slice-map.ts
@@ -2,11 +2,11 @@ import mapSlice from '../src/slice-map';
 
 // testing no params
 const one = mapSlice({ name: 'SLICE' });
-// $ExpectType { [x: string]: (payload?: any) => Action<any>; }
+// $ExpectType { add: (payload?: any) => Action<any>; set: (payload?: any) => Action<any>; remove: (payload: string[]) => Action<string[]>; patch: (payload: PatchEntity<AnyState>) => Action<PatchEntity<AnyState>>; reset: () => Action<...>; }
 one.actions;
 // $ExpectType Reducer<AnyState, Action<any>>
 one.reducer;
-// $ExpectType "SLICE"
+// $ExpectType string
 one.name;
 
 // testing with params
@@ -16,19 +16,11 @@ interface Obj {
 interface SliceState {
   [key: string]: Obj;
 }
-interface Actions {
-  add: SliceState;
-  set: SliceState;
-  remove: string[];
-  reset: never;
-}
-interface State {
-  slice: SliceState;
-}
-const two = mapSlice<SliceState, Actions, State>({ name: 'slice' });
-// $ExpectType { add: (payload: SliceState) => Action<SliceState>; set: (payload: SliceState) => Action<SliceState>; remove: (payload: string[]) => Action<string[]>; reset: () => Action<any>; }
+
+const two = mapSlice<SliceState>({ name: 'slice' });
+// $ExpectType { add: (payload: SliceState) => Action<SliceState>; set: (payload: SliceState) => Action<SliceState>; remove: (payload: string[]) => Action<string[]>; patch: (payload: PatchEntity<SliceState>) => Action<...>; reset: () => Action<...>; }
 two.actions;
 // $ExpectType Reducer<SliceState, Action<any>>
 two.reducer;
-// $ExpectType "slice"
+// $ExpectType string
 two.name;