+++ type-tests/slice.ts
@@ -1,22 +1,5 @@
 import createSlice from '../src/slice';
 
-// testing InputWithName and 3 params
-const name = 'something';
-interface ThreeState {
-  something: number;
-}
-const addThree = createSlice<number, { add: number }, ThreeState>({
-  name,
-  reducts: { add: (state, p: number) => state + p },
-  initialState: 0,
-});
-// $ExpectType { add: (payload: number) => Action<number>; }
-addThree.actions;
-// $ExpectType Reducer<number, Action<any>>
-addThree.reducer;
-// $ExpectType "something"
-addThree.name;
-
 // testing InputWithName and 2 params
 const addFour = createSlice<number, { add: number }>({
   name: 'something',
@@ -27,12 +10,12 @@ const addFour = createSlice<number, { add: number }>({
 addFour.actions;
 // $ExpectType Reducer<number, Action<any>>
 addFour.reducer;
-// $ExpectType string | number | symbol
+// $ExpectType string
 addFour.name;
 
 // testing InputWithName and no params
 const addFive = createSlice({
-  name: 'something',
+  name: 'something' as const,
   reducts: { add: (state, p: number) => state + p },
   initialState: 0,
 });
@@ -40,7 +23,7 @@ const addFive = createSlice({
 addFive.actions;
 // $ExpectType Reducer<number, Action<any>>
 addFive.reducer;
-// $ExpectType "something"
+// $ExpectType string
 addFive.name;
 
 // testing InputWithName initialized let
@@ -54,5 +37,5 @@ const addEight = createSlice({
 addEight.actions;
 // $ExpectType Reducer<number, Action<any>>
 addEight.reducer;
-// $ExpectType string | number
+// $ExpectType string
 addEight.name;